Improved todo page styles, fixed logout function, fixed preloader and error message handling, added success message, sort list by date ascending order, PROJECT V1 COMPLETED

+++ client/src/components/todo/ToDoItem.js
@@ -3,30 +3,46 @@ import moment from 'moment';
 import { deleteTodo, setCurrentTodo, clearCurrentTodo } from '../../actions/todoActions';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
+import { confirmAlert } from 'react-confirm-alert';
+import { setMessage } from '../../actions/messageActions';
 
-const ToDoItem = ({todo, deleteTodo, setCurrentTodo, clearCurrentTodo}) => {
+const ToDoItem = ({todo, deleteTodo, setCurrentTodo, clearCurrentTodo, setMessage}) => {
 
     const {_id, description, priority, date} = todo;
 
     const deleteThis = e => {
-        clearCurrentTodo();
-        deleteTodo(_id);
+        confirmAlert({
+            title: 'Permanently remove this?',
+            buttons: [
+                {
+                    label: 'Yes',
+                    onClick: () => {
+                        clearCurrentTodo();
+                        deleteTodo(_id);
+                    }
+                },
+                {
+                    label: 'Cancel'
+                }
+            ]
+        })
     }
 
     const setCurrent = e => {
+        setMessage("", "");
         setCurrentTodo(todo);
     }
 
     return (
         <Fragment>
-           <li className="collection-item flow-text">
+           <li className="collection-item flow-text row">
                 {priority === 'High' && (<span className="custom-badge high-priority-badge">H</span>)  }
                 {priority === 'Medium' && (<span className="custom-badge medium-priority-badge">M</span>)  }
                 {priority === 'Low' && (<span className="custom-badge low-priority-badge">L</span>)  }
-                {description}{(date !== null && date !== '') && ( <span> is to be done by {moment(date).format('DD-MM-YYYY')}</span>)}
+                <span>{description}</span>{(date !== null && date !== '') && ( <span> is to be done by <span className="date-text">{moment(date).format('DD-MM-YYYY')}</span></span>)}
                 <span className="secondary-content">
-                    <a href="#formModal" type="button" className="waves-effect waves-light btn modal-trigger edit-button" onClick={setCurrent}><i className="material-icons">edit</i></a>
-                    <button type="button" className="waves-effect waves-light btn delete-button" onClick={deleteThis}><i className="material-icons">delete_forever</i></button>
+                    <a href="#formModal" type="button" className="btn btn-primary modal-trigger edit-button" onClick={setCurrent}><i className="material-icons">edit</i></a>
+                    <button type="button" className="btn btn-secondary delete-button" onClick={deleteThis}><i className="material-icons">delete_forever</i></button>
                 </span>
             </li>
         </Fragment>
@@ -35,7 +51,9 @@ const ToDoItem = ({todo, deleteTodo, setCurrentTodo, clearCurrentTodo}) => {
 
 ToDoItem.propTypes = {
     deleteTodo: PropTypes.func.isRequired,
-    setCurrentTodo: PropTypes.func.isRequired
+    setCurrentTodo: PropTypes.func.isRequired,
+    clearCurrentTodo: PropTypes.func.isRequired,
+    setMessage: PropTypes.func.isRequired
 }
 
-export default connect(null, {deleteTodo, setCurrentTodo, clearCurrentTodo})(ToDoItem);
+export default connect(null, {deleteTodo, setCurrentTodo, clearCurrentTodo, setMessage})(ToDoItem);
